refatorado layout do carrinho, com componentização
